Part 13 of adding two new ctk configuration macros: - CTK_CONF_ICON_BITMAPS - CTK_CONF_ICON_TEXTMAPS
These macros control if icon bitmaps resp. textmaps are used at all. For all platfroms only needing one or the other this saves memory as the application description files don't define the unused stuff anymore. Especially for the conio based platforms this is a significant benefit.

Note: For all platforms (apart for Apple2 and Atari) I initially defined both macros as 1 which should lead to an unchanged bevaviour. I leave it up to the platform maintainers to make use of the new macros by optionally setting one to 0.