The mousetext based ctk drawing code uses the cc65 C library function _textframexy to draw the window borders.
This function doesn't support clipping, which results in minor screen corruption when clipping is actually performed. This is (at least currently) only the case for the redraw due to a menu close.

Therefore I opted to completely remove clipping support from the mousetext based drawing code which means that on each menu close the whole screen gets redrawn, not only the upper part of the screen down to the last line covered by the menu.

Given the quite high general screen drawing speed (due to working in text mode) this disadvantage seems to be less significant than the two advantages:

- No more screen corruption
- Saving 435 bytes (!) of code