On plaforms with (sub-)directory support it makes sense to seperate user file access (to be directed in to the current directory) from system file (.prg, .drv, ...) access (to be directed to the Contiki directory).
This is archivable even if there's no explicit path support in Contiki because all system file access goes through the loader-arch files, which may modify the file name given as they see fit - i.e. prepend it with an absolute path prefix.

The only area where this doesn't work out is directory listing. Both user file and system directory access uses cfs_opendir straight. Therefore different cfs_opendir parameters are the obvious solution.

The most widely adopted way to denote the current directory is ".". So I changed "/" to "." for all accesses to the current directory.

Note that up to now this was no issue as there was no platform with both subdirectories *AND* system files - which has changed recently with for the Apple2.