Added a command line argument to the apple2enh target.
One command line argument is accpeted which is the name of the application to be started initially - basically instead of 'welcome.prg'. The directory part is removed because application can only be loaded from the Contiki directory.

The intended usecase is selecting a Contiki *.prg file in the Apple GS/OS Finder with an Icon File in place for *.prg files which contains in the 'boss' section a reference to my ProDOS binary file loader. The loader supports the ProDOS startup file protocol used by the ProDOS launcher of GS/OS and emulates a BASIC REM command line understood by the cc65 C-Library command line handler.