The general policy of the configedit app is to _not_ load the configuration file for providing default values but to get them from the "live" system.
This gets a little complicated with the current directory value as it allows relative paths to be used - given to chdir() - while the system only returns the absolute path - taken from getcwd().

The approach choosen here is to store the length n of the current directoy value and to assuem that the last n characters of the absolute path are identical to that value - given that chdir() was successful at all. Otherwise the value is assumed to be invalid and thus is emptied.